+++ src/test/java/org/morphix/reflection/MethodsIgnoreAccessInvokeTest.java
@@ -26,7 +26,7 @@ import org.morphix.reflection.testdata.A;
  *
  * @author Radu Sebastian LAZIN
  */
-class MethodsIgnoreAccessInvokeMethodTest {
+class MethodsIgnoreAccessInvokeTest {
 
 	private static final String TEST_STRING = "Test";