+++ src/test/java/org/morphix/reflection/MethodsIgnoreAccessInvokeWithOriginalExceptionTest.java
@@ -26,7 +26,7 @@ import org.morphix.reflection.testdata.A;
  *
  * @author Radu Sebastian LAZIN
  */
-class MethodsIgnoreAccessInvokeMethodWithOriginalExceptionTest {
+class MethodsIgnoreAccessInvokeWithOriginalExceptionTest {
 
 	private static final String TEST_STRING = "Test";